GDA: Add custom card for Generative Direct Answers
This PR contains all the logic necessary to link the theme with
answers-search-ui for GDA. It also includes a custom card for GDA
in the theme (which looks essentially identical to the default).

To use GDA, the implementer will need to:
- uncomment the property in the global config json file
- uncomment the gda related lines in the page.html.hbs file(s)
- (if they want to use the theme's card) uncomment the GDA section
of the page-config json file(s)

J=WAT-4595, WAT-4599
TEST=manual
Spun up local test site, saw expected behavior. See demo here:
https://drive.google.com/file/d/1Bvi9TT5b1DYo84TZMxaeZDBXgqNtbqBk/view?usp=sharing